fix(ci): Sửa lỗi ESLint cho workflow của frontend
